converted int to short int

+++ getBreath1/include/Date.h
@@ -7,7 +7,7 @@ class Date {
 	short int year;
 public:
 	Date() :month{ 1 }, day{ 1 }, year{ 2000 } {};//default constructor
-	Date(int month, int day, int year);
+	Date(short int month, short int day, short int year);
 
 	friend inline std::istream& operator>>(istream& input, Date& date) { // will manipulate how we recive input
 		char slash;
